(almost) complete description of the new sout architecture, and of the available modules.
It is now integrated with the vlc-user-guide. It still has to be
integrated in the videolan-howto.

I propose not to integrate this complete explanation in the VH, but,
for each example, give, if possible, both syntaxes (the "old --sout
access/mux:url" and the new chained one), and give reference to the
VLC-UG for complete explanations.