*experimental* translation mechanism using po4a
po4a is "PO for anything". It allows to use PO files for virtually anything (using plugins).

po4a extracts strings from the english files, generates the pot files and updates the po files.

It then generates the translated documents in original format.


To update PO files from english and generate translated documents, use, in src/ :

"po4a po4a.cfg -k 0" (-k 0 tells it to generate translated documents
even if there is only 0% of it translated. Default threshold is 80%)


As I said, this is very experimental. If it is not satisfactory, we'll
get back to the previous translation system, but I think this one has
loads of advantages :

- easy tracking of modifications
- easy editing (kbabel, gtranslator, poedit, ...)
- reuse of strings when they appear several times