Skin editor: don't link to outdated and forgotten development page.
The project is dead, and the developer has no interest in it since a few
years.  He also does not have access to the developer pages, so he can't
fix the links there.

All downloads on that link are 403, so I guess our best way is to link
to officially released versions.